# api/openapi.yaml
openapi: 3.0.3
info:
  title: Calendar API
  version: 1.0.0
  contact:
    name: Source code
    url: https://github.com/Alexander-D-Karpov/calendar
  description: |
    JSON API for calendars, events, todos and share links. All operations live under `/api/v1`.

    ## Authentication

    Send an API token in the `Authorization` header:

        Authorization: Bearer cal_<prefix>_<secret>

    Create tokens in [Settings → API tokens](/settings/tokens). A token is shown once, right after it is created, and can be revoked at any time. To try requests on this page with a token, press **Authorize** and paste it.

    Requests from a signed-in browser session work too. With a session, unsafe methods (`POST`, `PUT`, `PATCH` and `DELETE`) also need the `X-CSRF-Token` header. **Try it out** on this page adds that header when you are signed in.

    ## Scopes

    Each token carries one or more scopes, and each operation names the scope it needs. A `write` scope includes `read` access to the same resource. Browser sessions have every scope.

    | Scope | Allows |
    |---|---|
    | `account:read` | Reading the account profile |
    | `account:write` | Changing account settings and the sleep schedule |
    | `calendars:read` | Reading calendars and events |
    | `calendars:write` | Creating, changing and deleting calendars and events |
    | `todos:read` | Reading todo lists, todos and checks |
    | `todos:write` | Creating, changing, completing and deleting todos |
...

...

    ## MCP

...

...
servers:
...
paths:
...
  /events:
    get:
      tags: [Events]
      operationId: listEvents
      summary: List events in a time range
      description: |
        Returns events that overlap the range from `from` to `to`, ordered by start. The range can be at most 400 days. All-day events are matched by date in the account timezone.

        Recurring series are expanded into occurrences. An occurrence of a series carries the series `id` and its original start in `instance`. An occurrence changed on its own has its own `id`, the series in `series_id` and its original start in `instance`. With `expand=false` the stored series, changed occurrences and single events are returned without expansion.

        Requires scope `calendars:read`.
      x-scope: calendars:read
      security: *readAuth
      parameters:
        - $ref: "#/components/parameters/RangeFrom"
        - $ref: "#/components/parameters/RangeTo"
        - $ref: "#/components/parameters/CalendarFilter"
        - $ref: "#/components/parameters/Expand"
      responses:
        "200":
          description: Events and occurrences in the range.
          headers: *okHeaders
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventList"
              example:
                items:
                  - &eventExample
                    id: "0191f2c5-1b2c-7d4e-8f90-a1b2c3d4e5f6"
                    calendar_id: "0191f2c4-7a1b-7c3e-9a51-2b7d0e4f6a10"
                    uid: "0191f2c5-1b2c-7d4e-8f90-a1b2c3d4e5f6"
                    series_id: null
                    instance: "2026-09-11T10:00:00+03:00"
                    recurring: true
                    title: Standup
                    body: "Agenda and **notes**"
                    location: Office
                    url: ""
                    all_day: false
                    start: "2026-09-11T10:00:00+03:00"
                    end: "2026-09-11T10:15:00+03:00"
                    timezone: Europe/Moscow
                    rrule: "FREQ=WEEKLY;BYDAY=MO,TU,WE,TH,FR"
                    rdate: []
                    exdate: ["2026-09-14T10:00:00+03:00"]
                    status: confirmed
                    transparency: opaque
                    visibility: default
                    color: null
                    read_only: false
                    reminders: [10]
                    version: 3
                    created_at: "2026-09-10T13:56:42Z"
                    updated_at: "2026-09-10T14:02:11Z"
                  - id: "0191f2c6-2d3e-7f40-8a51-b2c3d4e5f607"
                    calendar_id: "0191f2c4-7a1b-7c3e-9a51-2b7d0e4f6a10"
                    uid: "0191f2c6-2d3e-7f40-8a51-b2c3d4e5f607"
                    series_id: null
                    instance: null
                    recurring: false
                    title: Conference
                    body: ""
                    location: Vilnius
                    url: "https://example.com/conf"
                    all_day: true
                    start: "2026-09-20"
                    end: "2026-09-22"
                    timezone: null
                    rrule: null
                    rdate: []
                    exdate: []
                    status: tentative
                    transparency: transparent
                    visibility: private
                    color: "#aa3355"
                    read_only: false
                    reminders: []
                    version: 1
                    created_at: "2026-09-10T13:56:42Z"
                    updated_at: "2026-09-10T13:56:42Z"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
        "429":
          $ref: "#/components/responses/TooManyRequests"
    post:
      tags: [Events]
      operationId: createEvent
      summary: Create an event
      description: |
        Creates an event. Only `start` is required. Without `calendar_id` the event goes to the default calendar. Without `end` a timed event lasts one hour and an all-day event one day. Times without an offset are read in `timezone`, which defaults to the calendar timezone and then the account timezone. Without `reminders` the calendar's default reminders are used.

        Requires scope `calendars:write`.
      x-scope: calendars:write
      security: *writeAuth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EventInput"
            examples:
              recurring:
                summary: Weekday standup
                value:
                  title: Standup
                  start: "2026-09-11T10:00:00+03:00"
                  end: "2026-09-11T10:15:00+03:00"
                  timezone: Europe/Moscow
                  rrule: "FREQ=WEEKLY;BYDAY=MO,TU,WE,TH,FR"
                  reminders: [10]
              allDay:
                summary: All-day event
                value:
                  title: Conference
                  all_day: true
                  start: "2026-09-20"
                  end: "2026-09-22"
      responses:
        "201":
          description: The created event.
          headers:
            ETag:
              $ref: "#/components/headers/ETag"
            Location:
              $ref: "#/components/headers/Location"
            RateLimit-Limit:
              $ref: "#/components/headers/RateLimitLimit"
            RateLimit-Remaining:
              $ref: "#/components/headers/RateLimitRemaining"
            RateLimit-Reset:
              $ref: "#/components/headers/RateLimitReset"
            X-Request-ID:
              $ref: "#/components/headers/RequestID"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
              example: *eventExample
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "415":
          $ref: "#/components/responses/UnsupportedMediaType"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
        "429":
          $ref: "#/components/responses/TooManyRequests"
  /events/{id}:
    parameters:
      - $ref: "#/components/parameters/EventID"
    get:
      tags: [Events]
      operationId: getEvent
      summary: Get an event
      description: |
        Returns one stored event: a single event, a recurring series or an occurrence changed on its own. Use `GET /events/{id}/instances` for the occurrences of a series.

        Requires scope `calendars:read`.
      x-scope: calendars:read
      security: *readAuth
      parameters:
        - $ref: "#/components/parameters/IfNoneMatch"
      responses:
        "200":
          description: The event.
          headers: *resourceHeaders
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
              example: *eventExample
        "304":
          $ref: "#/components/responses/NotModified"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/TooManyRequests"
    patch:
      tags: [Events]
      operationId: updateEvent
      summary: Update an event
      description: |
        Changes the fields present in the body as a JSON Merge Patch. Changing `start` without `end` keeps the duration.

        For a recurring series, `scope` and `instance` choose what changes:

        - `this` changes one occurrence and returns it as a changed occurrence.
        - `following` ends the series before the occurrence and starts a new series there with the changes applied. The response is the new series. Its `uid` is the old one with `_R` and the occurrence start in UTC appended, as in `<uid>_R20260911T100000Z`, or a fresh ID when that would pass 1000 characters. Changed occurrences, extra dates and excluded dates at or after the split move to the new series.
        - `all` applies the fields to the series itself, so `start` sets the first occurrence. Moving the start also moves excluded dates, extra dates and changed occurrences by the same amount, unless the request sets `rdate` or `exdate` itself, or changes `all_day` or `rrule`.

        A changed occurrence that no longer falls on the rule is deleted.

        Recurrence fields can only change with `following` or `all`. `uid` cannot be changed.

        Requires scope `calendars:write`.
      x-scope: calendars:write
      security: *writeAuth
      parameters:
        - $ref: "#/components/parameters/EditScope"
        - $ref: "#/components/parameters/Instance"
        - $ref: "#/components/parameters/IfMatch"
      requestBody:
        required: true
        content:
          application/merge-patch+json:
            schema:
              $ref: "#/components/schemas/EventInput"
            example:
              start: "2026-09-11T11:00:00+03:00"
          application/json:
            schema:
              $ref: "#/components/schemas/EventInput"
            example:
              title: Standup moved
              location: null
      responses:
        "200":
          description: The changed event, changed occurrence or new series.
          headers: *resourceHeaders
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
              example: *eventExample
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "412":
          $ref: "#/components/responses/PreconditionFailed"
        "415":
          $ref: "#/components/responses/UnsupportedMediaType"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
        "429":
          $ref: "#/components/responses/TooManyRequests"
    delete:
      tags: [Events]
      operationId: deleteEvent
      summary: Delete an event
      description: |
        Deletes the event and moves it to the trash. For a recurring series, `this` excludes one occurrence and deletes the changed copy of it if there is one, `following` ends the series before the occurrence, and `all` deletes the series with every changed occurrence.

        Requires scope `calendars:write`.
      x-scope: calendars:write
      security: *writeAuth
      parameters:
        - $ref: "#/components/parameters/EditScope"
        - $ref: "#/components/parameters/Instance"
        - $ref: "#/components/parameters/IfMatch"
      responses:
        "204":
          description: The event or occurrences were deleted.
          headers: *okHeaders
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "412":
          $ref: "#/components/responses/PreconditionFailed"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
        "429":
          $ref: "#/components/responses/TooManyRequests"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: cal_<prefix>_<secret>
      description: API token created in Settings → API tokens.
    sessionCookie:
      type: apiKey
      in: cookie
      name: __Host-sid
      description: >-
        Browser session cookie, set by signing in. It is named `sid` when the server runs over plain HTTP.
        Unsafe methods also need the `X-CSRF-Token` header.
    csrfToken:
      type: apiKey
      in: header
      name: X-CSRF-Token
      description: >-
        CSRF token of the browser session, required together with `sessionCookie` on unsafe methods.
        Pages expose it in the `csrf-token` meta tag.
  parameters:
...
    RangeFrom:
      name: from
      in: query
      required: true
      description: Start of the range, inclusive. An RFC 3339 date-time, or a date for midnight in the account timezone.
      schema:
        type: string
      example: "2026-09-07T00:00:00+03:00"
    RangeTo:
      name: to
      in: query
      required: true
      description: End of the range, exclusive. An RFC 3339 date-time, or a date for midnight in the account timezone.
      schema:
        type: string
      example: "2026-09-14"
    CalendarFilter:
      name: calendar_id
      in: query
      required: false
      description: Only include these calendars. Repeat the parameter or separate IDs with commas. Without it, every calendar that is not hidden is included.
      style: form
      explode: true
      schema:
        type: array
        items:
          type: string
          format: uuid
      example: ["0191f2c4-7a1b-7c3e-9a51-2b7d0e4f6a10"]
    Expand:
      name: expand
      in: query
      required: false
      description: Expand recurring series into occurrences.
      schema:
        type: boolean
        default: true
      example: true
...
  headers:
    ETag:
      description: Version of the resource. Send it in `If-Match` to avoid overwriting other changes.
      schema:
        type: string
      example: '"lx3kq9zb"'
    Location:
      description: URL of the created resource.
      schema:
        type: string
      example: /api/v1/calendars/0191f2c4-7a1b-7c3e-9a51-2b7d0e4f6a10
...
  responses:
...
    Unauthorized:
      description: The token or session is missing, invalid, revoked or expired.
      headers:
        WWW-Authenticate:
          description: Authentication scheme to use.
          schema:
            type: string
          example: 'Bearer realm="calendar"'
        X-Request-ID:
          $ref: "#/components/headers/RequestID"
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
          example:
            type: "about:blank"
            title: Unauthorized
            status: 401
            detail: invalid api token
            instance: /api/v1/me
            request_id: tQ6DNZRbQcyrEPPA
    Forbidden:
      description: The token lacks the required scope, or a session request failed the CSRF or origin check.
      headers:
        X-Request-ID:
          $ref: "#/components/headers/RequestID"
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
          examples:
            scope:
              summary: Missing scope
              value:
                type: "about:blank"
                title: Forbidden
                status: 403
                detail: "token lacks scope calendars:write"
                instance: /api/v1/calendars
                request_id: tQ6DNZRbQcyrEPPA
            csrf:
              summary: Missing CSRF token
              value:
                type: "about:blank"
                title: Forbidden
                status: 403
                detail: invalid csrf token
                instance: /api/v1/calendars
                request_id: tQ6DNZRbQcyrEPPA
...
    UnprocessableEntity:
      description: The request body is well formed but some fields are invalid.
      headers:
        X-Request-ID:
          $ref: "#/components/headers/RequestID"
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
          example:
            type: "about:blank"
            title: Unprocessable Entity
            status: 422
            detail: One or more fields are invalid.
            instance: /api/v1/calendars
            request_id: tQ6DNZRbQcyrEPPA
            errors:
              - field: color
                message: "must be a hex color like #3b6ea5"
...
  schemas:
...
    EventList:
      type: object
      description: Events ordered by start.
      additionalProperties: false
      required: [items]
      properties:
        items:
          type: array
          description: Events and occurrences.
          items:
            $ref: "#/components/schemas/Event"
...